+++ packages/expo-dev-menu/vendored/react-native-reanimated/Common/cpp/Tools/Mapper.cpp
@@ -1,6 +1,6 @@
-#include "Mapper.h"
-#include "MutableValue.h"
-#include "SharedParent.h"
+#include "DevMenuMapper.h"
+#include "DevMenuMutableValue.h"
+#include "DevMenuSharedParent.h"
 
 namespace devmenureanimated {