+++ packages/expo-dev-menu/vendored/react-native-reanimated/ios/native/DevMenuREAIOSScheduler.h
@@ -1,4 +1,4 @@
-#import "Scheduler.h"
+#import "DevMenuScheduler.h"
 #import <React/RCTUIManager.h>
 #import <ReactCommon/CallInvoker.h>
 #include <stdio.h>