+++ packages/expo-dev-menu/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 🐛 Bug fixes
 
+- Fixed compatibility with the `react-native-reanimated` on iOS.
+
 ### 💡 Others
 
 ## 1.1.0 — 2022-07-18